Refactored project to use new webpack build
* Added new webpack configuration for umd build
* Added new es/umd build and npm publishing scripts to package.json
* Added umd build files and lib folder to .gitignore
* Removed build folder from .gitignore
* Moved source files into new src folder
* Added .npmignore to prevent src directory being included in package
* Refactored output path and filename in webpack test configuration
* Updated enzyme package and added react-test-renderer to remove warning
* Refactored page visibility to use es6 class to remove warnings
* Set react as a peer dependancy and dev dependancy so not included in
parent project webpack builds
* Added PropTypes package and refactored component
* Refactored eslint to no longer error on react/prefer-es6-class
