+++ .babelrc
@@ -1,3 +1,3 @@
 {
-  "presets": ["es2015", "react"]
+  "presets": [["es2015", { "modules": false }], "react"]
 }
