+++ .eslintrc
@@ -50,7 +50,6 @@
         "react/no-did-mount-set-state": 2,
         "react/no-did-update-set-state": 2,
         "react/no-direct-mutation-state": 2,
-        "react/prefer-es6-class": [2, "never"],
         "react/jsx-no-undef": 1,
         "react/jsx-uses-vars": 1,
         "react/prop-types": [0],
