+++ package.json
@@ -5,7 +5,8 @@
   "author": "Gilad Peleg <giladp007@gmail.com> (http://giladpeleg.com)",
   "repository": "pgilad/react-page-visibility",
   "keywords": "react, components, visibility, page, higher-order, browser",
-  "main": "index.js",
+  "main": "react-page-visibility",
+  "module": "lib/index",
   "files": [
     "index.js",
     "visibility.js"
@@ -14,15 +15,23 @@
     "node": ">=6.0.0"
   },
   "scripts": {
-    "build": "webpack",
-    "test": "tape -r ./test/babel-register-setup test/*-test.js | faucet",
-    "test-browser": "npm run build && cat ./build/bundle.js | tape-run -r faucet",
-    "test-ci": "npm run build && zuul --no-coverage --browser-output-timeout 15000 --concurrency 1 -- ./build/bundle.js",
-    "test-local": "npm run build && zuul --local -- ./build/bundle.js",
-    "test-phantom": "npm run build && zuul --phantom -- ./build/bundle.js"
+    "build": "npm run build:es && npm run build:umd ",
+    "build:es": "babel src --out-dir lib/ --source-maps",
+    "build:test": "node_modules/.bin/webpack --config webpack.test.config.js",
+    "build:umd": "node_modules/.bin/webpack",
+    "lint": "node_modules/.bin/eslint src",
+    "prepublish": "npm run build",
+    "preversion": "npm run test",
+    "test": "npm run build:test && node_modules/.bin/tape -r ./test/babel-register-setup test/*-test.js | node_modules/.bin/faucet",
+    "test-browser": "npm run build:test && cat ./react-page-visibility.test.js | node_modules/.bin/tape-run -r node_modules/.bin/faucet",
+    "test-ci": "npm run build:test && node_modules/.bin/zuul --no-coverage --browser-output-timeout 15000 --concurrency 1 -- ./react-page-visibility.test.js",
+    "test-local": "npm run build:test && node_modules/.bin/zuul --local -- ./react-page-visibility.test.js",
+    "test-phantom": "npm run build:test && node_modules/.bin/zuul --phantom -- ./react-page-visibility.test.js",
+    "version": "npm run build"
   },
   "license": "MIT",
   "devDependencies": {
+    "babel-cli": "^6.24.1",
     "babel-core": "^6.7.7",
     "babel-loader": "^6.2.4",
     "babel-preset-es2015": "^6.6.0",
@@ -30,21 +39,26 @@
     "babel-preset-react": "^6.5.0",
     "babel-register": "^6.7.2",
     "babel-tape-runner": "^2.0.1",
-    "enzyme": "^2.2.0",
+    "enzyme": "^2.8.2",
     "eslint": "^3.5.0",
     "eslint-plugin-react": "^6.2.2",
     "expect.js": "^0.3.1",
     "faucet": "0.0.1",
     "jsdom": "^9.5.0",
     "phantomjs-prebuilt": "^2.1.7",
+    "react": "^15.0.1",
     "react-addons-test-utils": "^15.0.1",
     "react-dom": "^15.0.1",
+    "react-test-renderer": "^15.5.4",
     "tape": "^4.5.1",
     "tape-run": "^2.1.3",
     "webpack": "^1.13.0",
     "zuul": "^3.10.1"
   },
-  "dependencies": {
+  "peerDependencies": {
     "react": "^15.0.1"
+  },
+  "dependencies": {
+    "prop-types": "^15.5.10"
   }
 }
