+++ src/visibility.js
